Refine session footer settings components
